Refactor Timeline link routes
After careful consideration it became clear that the code here was too
complicated. Previously, the timeline id was passed down and filtered in
a separate `<Timeline />` component. As a result, this created a duplicate
component with similar functionality to the `<All />` component.

By filtering a list of posts we can just pass the new list to the `<All />`
component and skip all of the duplication.

Note: The `<All />` component should likely be renamed.