Move post refresh to useEffect based on uid
Moving this component to a different more specific sideEffect allows for
better control of enabling and disabling the child listener. This also
fixes an issue where the child listener is attached when the client cannot
fetch posts such as first login. As a result, this caused errors and prevents
posts from loading initially.