Refactor resizeImage function
This function now returns an object with sizes and file. Previously, this
structure was not as organized and only returned a new file which made it
impossible to get important data such as image width and height.

This refactor also takes advantage of getBufferAsync to simplify function
chaning making the code more readable.