Create alias for firebase
While determining if this alias was necessary a comparison was performed
on importing the project initailized firebase from "./src/firebase" vs
"firebase" package. The comparison shows that package includes all of the
firebase imports instead of just the ones we're using in this project.
As a result, the alias for our config should be used to only import what
is necessary.